[fix] int to size_t

+++ 0x12-singly_linked_lists/0-print_list.c
@@ -11,7 +11,7 @@
 
 size_t print_list(const list_t *h)
 {
-	int lcount = 0;
+	size_t lcount = 0;
 
 	while (h != NULL)
 	{
